split classes into separate files per .NET conventions
