+++ TractorMod/TractorMod.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
-using PhthaloBlue;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewValley;
@@ -11,245 +10,11 @@ using StardewValley.Buildings;
 using StardewValley.Characters;
 using StardewValley.TerrainFeatures;
 using StardewValley.Tools;
+using TractorMod.Framework;
 using SFarmer = StardewValley.Farmer;
 
 namespace TractorMod
 {
-    public class TractorHouse : Building
-    {
-        public TractorHouse() : base()
-        {
-            buildingType = "Tractor House";
-            humanDoor = new Point(-1, -1);
-            animalDoor = new Point(-2, -1);
-            indoors = null;
-            nameOfIndoors = "";
-            baseNameOfIndoors = "";
-            nameOfIndoorsWithoutUnique = "";
-            magical = false;
-            tileX = 0;
-            tileY = 0;
-            maxOccupants = 0;
-            tilesWide = 4;
-            tilesHigh = 2;
-            texture = Game1.content.Load<Texture2D>("..\\Mods\\TractorMod\\assets\\TractorHouse");
-            daysOfConstructionLeft = 1;
-        }
-
-        public TractorHouse(BluePrint input, Vector2 tileLocation) : base(input, tileLocation)
-        {
-            buildingType = "Tractor House";
-            humanDoor = new Point(-1, -1);
-            animalDoor = new Point(-2, -1);
-            indoors = null;
-            nameOfIndoors = "";
-            baseNameOfIndoors = "";
-            nameOfIndoorsWithoutUnique = "";
-            magical = true;
-            tileX = (int)tileLocation.X;
-            tileY = (int)tileLocation.Y;
-            maxOccupants = 0;
-            tilesWide = 4;
-            tilesHigh = 2;
-            texture = Game1.content.Load<Texture2D>("..\\Mods\\TractorMod\\assets\\Stable");
-            daysOfConstructionLeft = 0;
-        }
-
-        public TractorHouse(Vector2 tileLocation) : base()
-        {
-            buildingType = "Tractor House";
-            humanDoor = new Point(-1, -1);
-            animalDoor = new Point(-2, -1);
-            indoors = null;
-            nameOfIndoors = "";
-            baseNameOfIndoors = "";
-            nameOfIndoorsWithoutUnique = "";
-            magical = true;
-            tileX = (int)tileLocation.X;
-            tileY = (int)tileLocation.Y;
-            maxOccupants = 0;
-            tilesWide = 4;
-            tilesHigh = 2;
-            texture = Game1.content.Load<Texture2D>("..\\Mods\\TractorMod\\assets\\Stable");
-            daysOfConstructionLeft = 0;
-        }
-
-        public TractorHouse SetDaysOfConstructionLeft(int input)
-        {
-            daysOfConstructionLeft = input;
-            return this;
-        }
-
-        public override bool intersects(Rectangle boundingBox)
-        {
-            if (daysOfConstructionLeft > 0)
-                return base.intersects(boundingBox);
-            if (!base.intersects(boundingBox))
-                return false;
-            if (boundingBox.X >= (this.tileX + 1) * Game1.tileSize && boundingBox.Right < (this.tileX + 3) * Game1.tileSize)
-                return boundingBox.Y <= (this.tileY + 1) * Game1.tileSize;
-            return true;
-        }
-
-        public override void draw(SpriteBatch b)
-        {
-            if (this.daysOfConstructionLeft > 0)
-            {
-                this.drawInConstruction(b);
-            }
-            else
-            {
-                this.drawShadow(b, -1, -1);
-                b.Draw(this.texture, Game1.GlobalToLocal(Game1.viewport, new Vector2((float)(this.tileX * Game1.tileSize), (float)(this.tileY * Game1.tileSize + this.tilesHigh * Game1.tileSize))), new Rectangle?(this.texture.Bounds), this.color * this.alpha, 0.0f, new Vector2(0.0f, (float)this.texture.Bounds.Height), 4f, SpriteEffects.None, (float)((this.tileY + this.tilesHigh - 1) * Game1.tileSize) / 10000f);
-            }
-        }
-
-        public override void dayUpdate(int dayOfMonth)
-        {
-            base.dayUpdate(dayOfMonth);
-            if (this.daysOfConstructionLeft > 0)
-                return;
-            //this.grabHorse();
-            //do special action for this building here
-            /*
-             */
-        }
-
-        public override Rectangle getSourceRectForMenu()
-        {
-            return new Rectangle(0, 0, this.texture.Bounds.Width, this.texture.Bounds.Height);
-        }
-    }
-
-    public class Tractor : Horse
-    {
-        public Tractor() : base() { }
-        public Tractor(int tileX, int tileY) : base(tileX, tileY)
-        {
-            this.sprite = new AnimatedSprite(Game1.content.Load<Texture2D>("..\\Mods\\TractorMod\\assets\\tractor"), 0, 32, 32);
-            this.sprite.textureUsesFlippedRightForLeft = true;
-            this.sprite.loop = true;
-            this.faceDirection(3);
-        }
-        public override Rectangle GetBoundingBox()
-        {
-            Rectangle boundingBox = base.GetBoundingBox();
-            if ((this.facingDirection == 0 || this.facingDirection == 2))
-                boundingBox.Inflate(-Game1.tileSize / 2 - Game1.pixelZoom, 0);
-            return boundingBox;
-        }
-    }
-
-    public class SaveCollection
-    {
-        public class Save
-        {
-            public string FarmerName { get; set; } = "";
-            public ulong SaveSeed { get; set; }
-            public List<Vector2> TractorHouse = new List<Vector2>();
-            public Save() { SaveSeed = ulong.MaxValue; }
-            public Save(string nameInput, ulong input)
-            {
-                SaveSeed = input;
-                FarmerName = nameInput;
-            }
-
-            public Save AddTractorHouse(int inputX, int inputY)
-            {
-                foreach (Vector2 THS in TractorHouse)
-                {
-                    if (THS.X == inputX && THS.Y == inputY)
-                        return this;
-                }
-                TractorHouse.Add(new Vector2(inputX, inputY));
-                return this;
-            }
-        }
-        public List<Save> saves = new List<Save>();
-        public SaveCollection() { }
-        public SaveCollection Add(Save input)
-        {
-            saves.Add(input);
-            return this;
-        }
-        public Save FindSave(string nameInput, ulong input)
-        {
-            foreach (SaveCollection.Save ASave in saves)
-            {
-                if (ASave.SaveSeed == input && ASave.FarmerName == nameInput)
-                {
-                    return ASave;
-                }
-            }
-            return new SaveCollection.Save();
-        }
-
-        public Save FindCurrentSave()
-        {
-            return FindSave(Game1.player.name, Game1.uniqueIDForThisGame);
-        }
-    }
-
-    public class TractorConfig
-    {
-        public class ToolConfig
-        {
-            public string name { get; set; } = "";
-            public int minLevel { get; set; } = 0;
-            public int effectRadius { get; set; } = 1;
-            public int activeEveryTickAmount { get; set; } = 1;
-            private int activeCount = 0;
-
-            public ToolConfig() { }
-
-            public ToolConfig(string nameInput)
-            {
-                name = nameInput;
-            }
-
-            public ToolConfig(string nameInput, int minLevelInput, int radiusInput, int activeEveryTickAmountInput)
-            {
-                name = nameInput;
-                minLevel = minLevelInput;
-                effectRadius = radiusInput;
-                if (activeEveryTickAmountInput <= 0)
-                    activeEveryTickAmount = 1;
-                else
-                    activeEveryTickAmount = activeEveryTickAmountInput;
-            }
-
-            public void incrementActiveCount()
-            {
-                activeCount++;
-                activeCount %= activeEveryTickAmount;
-            }
-
-            public bool canToolBeActive()
-            {
-                return (activeCount == 0);
-            }
-        }
-        public string info1 = "Add tool with exact name you would like to use with Tractor Mode.";
-        public string info2 = "Also custom minLevel and effective radius for each tool.";
-        public string info3 = "Ingame tools included: Pickaxe, Axe, Hoe, Watering Can.";
-        public string info4 = "I haven't tried tools like Shears or Milk Pail but you can :)";
-        public string info5 = "Delete Scythe entry if you don't want to harvest stuff.";
-        public ToolConfig[] tool { get; set; } = {
-                                                        new ToolConfig("Scythe", 0, 2, 1),
-                                                        new ToolConfig("Hoe"),
-                                                        new ToolConfig("Watering Can")
-                                                     };
-
-        public int ItemRadius { get; set; } = 1;
-        public int holdActivate { get; set; } = 0;
-        public Keys tractorKey { get; set; } = Keys.B;
-        public int tractorSpeed { get; set; } = -2;
-        public Keys horseKey { get; set; } = Keys.None;
-        public Keys PhoneKey { get; set; } = Keys.N;
-        public int TractorHousePrice { get; set; } = 150000;
-        public Keys updateConfig { get; set; } = Keys.P;
-    }
-
     public class TractorMod : Mod
     {
         static Vector2 tractorSpawnLocation = new Vector2(70, 13);
